Fail terminally when the export warehouse has no identity row
